backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5

numpy==1.26.4
pandas==2.3.3
scikit-learn==1.8.0
xgboost==1.7.6
lightgbm==4.6.0

networkx==3.6.1
node2vec==0.5.0

geoip2==5.2.0
gensim==4.4.0
imbalanced-learn==0.14.1

shap==0.49.1
matplotlib==3.10.8
scipy==1.17.0
numba==0.65.0